the "Snow" update
i'm sorry i have updated this a lot today, i'm just having a lot of fun! in this one i added the snow wand. when you right click it throws a snowball and "freezes" people around where it lands! this one also includes code for ammo. so whenever it shoots, it takes away ammo in the players inventory. and if the player runs out of inventory it plays a clicking sound like an empty gun in real life and sends the player a warning message! it also adds ice around the players! i am kinda new to java, and hope this works.